requirements.txt
appnope==0.1.3
asttokens==2.0.5
backcall==0.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
chart-studio==1.1.0
click==8.1.3
confluent-kafka==1.9.2
cycler==0.11.0
dash==2.7.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.0
decorator==5.1.1
entrypoints==0.4
executing==0.8.3
findspark==2.0.1
Flask==2.2.2
fonttools==4.33.3
idna==3.4
importlib-metadata==5.1.0
ipykernel==6.13.0
ipython==8.2.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jupyter-client==7.2.2
jupyter-core==4.10.0
kafka-python==2.0.2
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
nest-asyncio==1.5.5
numpy==1.22.3
opencv-python==4.6.0.66
packaging==21.3
pandas==1.4.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
plotly==5.11.0
prompt-toolkit==3.0.29
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.5
pybind11==2.9.2
Pygments==2.11.2
pyparsing==3.0.8
pyspark==3.3.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.1
pyzmq==22.3.0
requests==2.28.1
retrying==1.3.4
six==1.16.0
stack-data==0.2.0
style==1.1.0
tenacity==8.1.0
tornado==6.1
traitlets==5.1.1
update==0.0.1
urllib3==1.26.13
wcwidth==0.2.5
Werkzeug==2.2.2
zipp==3.11.0
